Channelize output

+++ too/exec.go
@@ -2,21 +2,32 @@ package too
 
 import (
 	"fmt"
+	"io"
 	"os"
 	"os/signal"
 )
 
 // Exec ...
-func Exec(commands ...*Command) error {
+func Exec(output io.Writer, commands ...*Command) error {
 
 	interrupt := make(chan os.Signal, 1)
+	defer close(interrupt)
 	signal.Notify(interrupt, os.Interrupt)
 
 	endups := make(chan *Command)
+	defer close(endups)
 	endcnt := 0
 
+	stdout := output
+	stderr := output
+	if stderr == os.Stdout {
+		stderr = os.Stderr
+	}
+	msg := make(chan Message)
+	go msgQueue(stdout, stderr, msg)
+
 	for _, c := range commands {
-		if err := c.Start(endups); err != nil {
+		if err := c.Start(msg, endups); err != nil {
 			return err
 		}
 	}
@@ -26,6 +37,10 @@ func Exec(commands ...*Command) error {
 		case _ = <-interrupt:
 			errors := []error{}
 			for _, c := range commands {
+				if c.Process.Pid < 0 {
+					// This process is already released
+					continue
+				}
 				if err := c.Process.Kill(); err != nil {
 					errors = append(errors, err)
 				}
@@ -33,15 +48,33 @@ func Exec(commands ...*Command) error {
 			if len(errors) != 0 {
 				fmt.Println(errors)
 			}
-			// os.Exit(0)
 			break
 		case c := <-endups:
 			endcnt++
 			c.Wait()
-			c.PrintExitCode()
+			msg <- c.ExitCode()
 			if endcnt >= len(commands) {
+				// To make sure to print the last message
+				msg <- Message{Output: AppEnd}
 				return nil
 			}
+		}
+	}
+}
+
+func msgQueue(stdout, stderr io.Writer, msgchan chan Message) {
+	for {
+		m := <-msgchan
+		switch m.Output {
+		case AppEnd:
+			close(msgchan)
+			return
+		case Stdout:
+			m.Color.Fprintf(stdout, m.Header)
+			fmt.Fprintf(stdout, "%s\n", m.Text)
+		case Stderr:
+			m.Color.Fprintf(stderr, m.Header)
+			fmt.Fprintf(stderr, "%s\n", m.Text)
 		}
 	}
 }
